[PAR-4830] delete on config disable / v1 patch wip
